Implmenting pdfkit and deleting files.

+++ controllers/admin.js
@@ -1,5 +1,6 @@
 const Product = require("../models/product");
 const { validationResult } = require("express-validator/check");
+const fileHelper = require("../util/file");
 
 exports.getAddProduct = (req, res, next) => {
   res.render("admin/edit-product", {
@@ -19,7 +20,7 @@ exports.postAddProduct = (req, res, next) => {
   const description = req.body.description;
   const price = req.body.price;
   const errors = validationResult(req);
-  if(!image){
+  if (!image) {
     return res.status(422).render("admin/edit-product", {
       pageTitle: "Add Product",
       path: "/admin/addproduct",
@@ -30,7 +31,7 @@ exports.postAddProduct = (req, res, next) => {
         description: description,
         price: price
       },
-      errorMessage: 'Attached file is not an image.',
+      errorMessage: "Attached file is not an image.",
       validationError: []
     });
   }
@@ -51,7 +52,7 @@ exports.postAddProduct = (req, res, next) => {
     });
   }
   console.log("In add");
-  const imageUrl='/'+image.path;
+  const imageUrl = "/" + image.path;
   req.user
     .createProduct({
       title: title,
@@ -63,10 +64,10 @@ exports.postAddProduct = (req, res, next) => {
       console.log("product created");
       res.redirect("/admin/products");
     })
-    .catch(err =>{
+    .catch(err => {
       console.log(err);
-      const error=new Error(err);
-      error.httpStatusCode=500;
+      const error = new Error(err);
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
@@ -92,8 +93,8 @@ exports.getEditProduct = (req, res, next) => {
         validationError: []
       });
     })
-    .catch(err =>{
-      const error=new Error(err);
+    .catch(err => {
+      const error = new Error(err);
       error.httpStatusCode(500);
       next(error);
     });
@@ -102,11 +103,10 @@ exports.getEditProduct = (req, res, next) => {
 exports.postEditProduct = (req, res, next) => {
   const id = req.body.id;
   const title = req.body.title;
-  const image= req.file;
+  const image = req.file;
   const description = req.body.description;
   const price = req.body.price;
   const errors = validationResult(req);
-
 
   if (!errors.isEmpty()) {
     return res.status(422).render("admin/edit-product", {
@@ -118,10 +118,10 @@ exports.postEditProduct = (req, res, next) => {
         title: title,
         description: description,
         price: price,
-        id:id
+        id: id
       },
       errorMessage: errors.array()[0].msg,
-      validationError:errors.array()
+      validationError: errors.array()
     });
   }
   Product.findByPk(id)
@@ -130,8 +130,9 @@ exports.postEditProduct = (req, res, next) => {
         return res.redirect("/");
       }
       product.title = title;
-      if(image){
-        product.imageUrl ='/'+image.path;
+      if (image) {
+        fileHelper(product.imageUrl);
+        product.imageUrl = "/" + image.path;
       }
       product.description = description;
       product.price = price;
@@ -139,10 +140,10 @@ exports.postEditProduct = (req, res, next) => {
         res.redirect("/admin/products");
       });
     })
-    .catch(err =>{
+    .catch(err => {
       console.log(err);
-      const error=new Error(err);
-      error.httpStatusCode=500;
+      const error = new Error(err);
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
@@ -157,27 +158,32 @@ exports.getProducts = (req, res, next) => {
         pageTitle: "All Products"
       });
     })
-    .catch(err =>{
-      const error=new Error(err);
-      error.httpStatusCode=500;
+    .catch(err => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
 
 exports.deleteProduct = (req, res, next) => {
   const id = req.body.productId;
-  Product.destroy({
-    where: {
-      id: productId,
-      userId: req.user.id
-    }
-  })
+  Product.findByPk(id)
+    .then(product => {
+      if (!product) return next(new Error("Product not found!"));
+      fileHelper(product.imageUrl);
+      return Product.destroy({
+        where: {
+          id: id,
+          userId: req.user.id
+        }
+      });
+    })
     .then(result => {
       res.redirect("/admin/products");
     })
-    .catch(err =>{
-      const error=new Error(err);
-      error.httpStatusCode=500;
+    .catch(err => {
+      const error = new Error(err);
+      error.httpStatusCode = 500;
       return next(error);
     });
 };
